+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Account Gold Tracker
-### [1.0.0](https://github.com/diomsg-code/AccountGoldTracker/tree/1.0.0) (@build-date@)
-[Full Changelog](https://github.com/diomsg-code/AccountGoldTracker/commits/1.0.0) - [Previous Releases](https://github.com/diomsg-code/AccountGoldTracker/releases)
+### [1.1.0](https://github.com/diomsg-code/AccountGoldTracker/tree/1.1.0) (@build-date@)
+[Full Changelog](https://github.com/diomsg-code/AccountGoldTracker/compare/1.0.0...1.1.0) - [Previous Releases](https://github.com/diomsg-code/AccountGoldTracker/releases)
 
-- Initial release
+- Added: The difference to the previous day is displayed in the gold overview
+- Changed: Tabular view of the gold overview
+- Changed: New addon name